adds logging support to the catalog api

+++ src/Services/Catalog/Catalog.API/Controllers/CatalogController.cs
@@ -1,10 +1,9 @@
 using Catalog.API.Entities;
 using Catalog.API.Repositories;
 using Microsoft.AspNetCore.Mvc;
-using System.Net;
 
-
-namespace Catalog.API.Controllers {
+namespace Catalog.API.Controllers
+{
   [ApiController]
   [Route("api/v1/[controller]")]
   public class CatalogController : ControllerBase
@@ -19,18 +18,28 @@ namespace Catalog.API.Controllers {
     }
 
     [HttpGet]
-    [ProducesResponseType(typeof(IEnumerable<Product>), (int)HttpStatusCode.OK)]
+    [ProducesResponseType(typeof(IEnumerable<Product>), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<IEnumerable<Product>>> GetProducts()
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       var products = await _repository.GetProducts();
       return Ok(products);
     }
 
     [HttpGet("{id:length(24)}", Name = "GetProduct")]
-    [ProducesResponseType((int)HttpStatusCode.NotFound)]
-    [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(typeof(Product), StatusCodes.Status200OK)]
     public async Task<ActionResult<Product>> GetProductById(string id)
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       var product = await _repository.GetProduct(id);
       if (product == null)
       {
@@ -42,32 +51,66 @@ namespace Catalog.API.Controllers {
 
     [Route("[action]/{category}", Name = "GetProductByCategory")]
     [HttpGet]
-    [ProducesResponseType(typeof(IEnumerable<Product>), (int)HttpStatusCode.OK)]
+    [ProducesResponseType(typeof(IEnumerable<Product>), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<IEnumerable<Product>>> GetProductByCategory(string category)
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       var products = await _repository.GetProductByCategory(category);
       return Ok(products);
     }
 
+    [Route("[action]", Name = "GetProductByName")]
+    [HttpGet]
+    [ProducesResponseType(typeof(IEnumerable<Product>), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    public async Task<ActionResult<IEnumerable<Product>>> GetProductByName([FromQuery] string name)
+    {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
+      var products = await _repository.GetProductByName(name);
+      return Ok(products);
+    }
+
     [HttpPost]
-    [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
+    [ProducesResponseType(typeof(Product), StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<Product>> CreateProduct([FromBody] Product product)
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       await _repository.CreateProduct(product);
       return CreatedAtRoute("GetProduct", new { id = product.Id }, product);
     }
 
     [HttpPut]
-    [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
+    [ProducesResponseType(typeof(Product), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<IActionResult> UpdateProduct(Product product)
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       return Ok(await _repository.UpdateProduct(product));
     }
 
     [HttpDelete("{id:length(24)}", Name = "DeleteProduct")]
-    [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
+    [ProducesResponseType(typeof(Product), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     public async Task<IActionResult> DeleteProductById(string id)
     {
+      if (!ModelState.IsValid)
+      {
+        return BadRequest(ModelState);
+      }
       return Ok(await _repository.DeleteProduct(id));
     }
   }
